Introducing the new field -- 'tag_time' -- in tag_map table.
This will record the time at which a given media entry was tagged. A
proper table field macro was created for it. SELECT_MEDIA_FROM_TAG, a
query that is used by media_tag_foreach_media_from_db(), was changed
in order to expose the tag field to the user now, so that she can
apply filters (such as order by) on it.

Change-Id: I8972fa68d1caffbec2289fc44c141d64fcd3e886
Signed-off-by: Gustavo Lima Chaves <gustavo.lima.chaves@intel.com>